requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.3.11
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autopep8==2.3.2
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
catboost==1.2.8
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cssselect==1.3.0
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
dicttoxml==1.7.16
dill==0.4.0
docutils==0.21.2
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.7.0
future==1.0.0
geocoder==1.38.1
geoip2==5.1.0
graphviz==0.21
h11==0.16.0
htmlmin==0.1.12
httpcore==1.0.9
httpx==0.28.1
id==1.5.0
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.13.0
ip2geotools==0.1.6
IP2Location==8.10.5
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isort==6.0.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keyring==25.6.0
kiwisolver==1.4.8
lazy-object-proxy==1.11.0
lightgbm==4.6.0
llvmlite==0.44.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
maxminddb==2.7.0
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.7.0
multidict==6.6.3
multimethod==1.12
narwhals==1.47.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nh3==0.2.22
notebook==7.4.4
notebook_shim==0.2.4
numba==0.61.0
numpy==2.1.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
phik==0.12.5
pillow==11.3.0
pip-review==1.3.0
pkginfo==1.12.1.2
platformdirs==4.3.8
plotly==6.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
puremagic==1.30
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pylint==3.3.7
pyparsing==3.2.3
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
PyWavelets==1.8.0
pywin32==311
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.0
ratelim==0.1.6
readme_renderer==44.0
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.26.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
selenium==4.34.2
Send2Trash==1.8.3
setuptools==80.9.0
shap==0.48.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.5
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
twine==6.1.0
typed_ast==1.5.5
typeguard==4.4.4
types-python-dateutil==2.9.0.20250708
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
visions==0.8.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xgboost==3.0.2
yarl==1.20.1
ydata-profiling==4.16.1
-e git+https://github.com/EYOELTEKLE/fraud-cases-for-e-commerce-and-bank-transactions.git@fe263d229b6a8c72b2e823cfa4c9d17bd5808a86#egg=your_project_name
